groups

+++ Website/manager/views.py
@@ -13,12 +13,34 @@ from random import randint
 
 def manager_list(request):
 
+    if not request.user.is_authenticated:
+        return redirect('mylogin')
+
+    perm = 0
+    for i in request.user.groups.all():
+        if i.name == "masteruser": perm = 1
+
+    if perm == 0:
+        error = "Access Denied"
+        return render(request, 'back/error.html', {'error': error})
+
     manager = Manager.objects.all()
 
     return render(request, 'back/manager_list.html',{'manager': manager})
 
 
 def manager_del(request, pk):
+
+    if not request.user.is_authenticated:
+        return redirect('mylogin')
+
+    perm = 0
+    for i in request.user.groups.all():
+        if i.name == "masteruser": perm = 1
+
+    if perm == 0:
+        error = "Access Denied"
+        return render(request, 'back/error.html', {'error': error})
 
     manager = Manager.objects.get(pk=pk)
     b = User.objects.filter(username=manager.utxt)
@@ -31,12 +53,34 @@ def manager_del(request, pk):
 
 def manager_group(request):
 
-    group = Group.objects.all()
+    if not request.user.is_authenticated:
+        return redirect('mylogin')
+
+    perm = 0
+    for i in request.user.groups.all():
+        if i.name == "masteruser": perm =1
+
+    if perm == 0:
+        error = "Access Denied"
+        return render(request, 'back/error.html', {'error': error})
+
+    group = Group.objects.all().exclude(name="masteruser")
 
     return render(request, 'back/manager_group.html', {'group': group})
 
 
 def manager_group_add(request):
+
+    if not request.user.is_authenticated:
+        return redirect('mylogin')
+
+    perm = 0
+    for i in request.user.groups.all():
+        if i.name == "masteruser": perm = 1
+
+    if perm == 0:
+        error = "Access Denied"
+        return render(request, 'back/error.html', {'error': error})
 
     if request.method == 'POST':
 
@@ -54,6 +98,17 @@ def manager_group_add(request):
 
 def manager_group_del(request,name):
 
+    if not request.user.is_authenticated:
+        return redirect('mylogin')
+
+    perm = 0
+    for i in request.user.groups.all():
+        if i.name == "masteruser": perm =1
+
+    if perm == 0:
+        error = "Access Denied"
+        return render(request, 'back/error.html', {'error': error})
+
     b = Group.objects.filter(name=name)
     b.delete()
 
@@ -62,12 +117,71 @@ def manager_group_del(request,name):
 
 def users_groups(request, pk):
 
+    if not request.user.is_authenticated:
+        return redirect('mylogin')
+
+    perm = 0
+    for i in request.user.groups.all():
+        if i.name == "masteruser": perm = 1
+
+    if perm == 0:
+        error = "Access Denied"
+        return render(request, 'back/error.html', {'error': error})
+
     manager = Manager.objects.get(pk=pk)
 
     user = User.objects.get(username=manager.utxt)
 
-    group = []
+    ugroup = []
     for i in user.groups.all():
-        group.append(i.name)
+        ugroup.append(i.name)
 
-    return render(request, 'back/users_groups.html', {'group': group})
+    group = Group.objects.all()
+
+    return render(request, 'back/users_groups.html', {'ugroup': ugroup, 'group': group, 'pk': pk})
+
+
+def add_users_to_groups(request, pk):
+
+    if not request.user.is_authenticated:
+        return redirect('mylogin')
+
+    perm = 0
+    for i in request.user.groups.all():
+        if i.name == "masteruser": perm = 1
+
+    if perm == 0:
+        error = "Access Denied"
+        return render(request, 'back/error.html', {'error': error})
+
+    if request.method == 'POST':
+
+        gname = request.POST.get('gname')
+
+        group = Group.objects.get(name=gname)
+        manager = Manager.objects.get(pk=pk)
+        user = User.objects.get(username=manager.utxt)
+        user.groups.add(group)
+
+    return redirect('users_groups', pk=pk)
+
+
+def del_users_to_groups(request, pk, name):
+
+    if not request.user.is_authenticated:
+        return redirect('mylogin')
+
+    perm = 0
+    for i in request.user.groups.all():
+        if i.name == "masteruser": perm = 1
+
+    if perm == 0:
+        error = "Access Denied"
+        return render(request, 'back/error.html', {'error': error})
+
+    group = Group.objects.get(name=name)
+    manager = Manager.objects.get(pk=pk)
+    user = User.objects.get(username=manager.utxt)
+    user.groups.remove(group)
+
+    return redirect('users_groups', pk=pk)
